x/toe/types/message_new_game_test.go
package types

import (
	"testing"

	"github.com/avikj/toe/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgNewGame_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgNewGame
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgNewGame{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgNewGame{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
